bike_rental/requirements.txt
asgiref==3.7.2
attrs==20.3.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
cryptography==41.0.3
distro==1.8.0
Django==4.2.7
django-debug-toolbar==4.2.0
docker==7.0.0
docker-composer-v2==0.8.8
dockerpty==0.4.1
docopt==0.6.2
factory-boy==3.3.0
Faker==20.1.0
idna==3.4
jsonschema==3.2.0
loguru==0.5.3
mbstrdecoder==1.1.3
numpy==1.26.1
packaging==23.1
pandas==2.1.2
paramiko==3.3.1
Pillow==10.1.0
psycopg==3.1.10
psycopg2==2.9.7
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
texttable==1.6.7
typepy==1.3.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
websocket-client==0.59.0
